splitting out parallel boundary code from other boundary code, checking for USE_MPI in platformEnv. Could also split out platformEnv into two classes
